Modified function of getting pivot
This modification changes the way of getting the pivot to median instead of
mean. This ensures more reliability in the correction of the project for
people that seek a specific case to fail it. The previous case was better in
the general case because getting the pivot was O(n) complexity, but was prone
to more moves with specific cases. The median has O(n^2) complexity but ensures
the moves stay constant. Median makes the overall program slower but the moves
stay low with any case.

+++ src/ft_get_input.c
@@ -6,7 +6,7 @@
 /*   By: orudek <orudek@student.42madrid.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/08/26 21:09:22 by orudek            #+#    #+#             */
-/*   Updated: 2023/09/15 15:37:28 by orudek           ###   ########.fr       */
+/*   Updated: 2023/09/15 20:48:38 by orudek           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,8 +30,8 @@ char	ft_is_number(char *str)
 
 static char	ft_is_too_long(char *str)
 {
-	long	num;
-	char	sign;
+	long long	num;
+	char		sign;
 
 	num = 0;
 	sign = 1;
